+++ api/api/spec/changes.md
@@ -20,17 +20,17 @@ a human readable message is shown, the new field `message` will be used instead.
 
 ### DELETE /agents
 * Parameter **ids** must be in query, not in body because DELETE operations can't have a requestBody in OpenAPI 3
-* Parameter **status** renamed to **agent_status**
 * In response, `msg` key is now moved to new `message` key
 
 ### GET /agents
-* Parameter **status** renamed to **agent_status**
 * Parameter **os.name** renamed to **os_name**
 * Parameter **os.platform** renamed to **os_platform**
 * Parameter **os.version** renamed to **os_version**
 
 ### GET /agents/groups/{group_id}
-* Parameter **status** renamed to **agent_status**
+
+### GET /agents/groups/{group_id}/configuration
+* In response, `filter` key is now moved to new `filters` key
 
 ### POST /agents/groups/{group_id}
 * In response, `msg` key is now moved to new `message` key
@@ -63,6 +63,9 @@ a human readable message is shown, the new field `message` will be used instead.
 
 ### DELETE /agents/groups/:group_id
 * In response, `msg` key is now moved to new `message` key
+
+### POST /agents/groups/:group_id
+* In response, now when group don't exists return a WazuhError and when agent don't exists return error infomation in failed_items section.
 
 ### PUT /agents/groups/:group_id
 * In response, `data` key is now moved to new `message` key